[Refactoring] Refactoring test.cpp file by making test.h
